+++ assets/javascripts/kitten/components/molecules/cards/clickable-card/index.js
@@ -58,7 +58,7 @@ const StyledClickableCard = styled.button`
   }
 
   .k-ClickableCard__image--circle {
-    border-radius: 50%;
+    border-radius: var(--border-radius-rounded);
   }
 
   .k-ClickableCard__content {