+++ assets/javascripts/kitten/components/organisms/stepper/index.js
@@ -64,7 +64,7 @@ const StepperText = styled(Text)`
   align-items: center;
   white-space: nowrap;
   transition: color 0.2s ease;
-  border-radius: ${pxToRem(6)};
+  border-radius: var(--border-radius-s);
 `
 
 export const StepperItem = ({ children, state, ...props }) => {